requirements.txt
pandas
jupyter
lxml
spacy
pycountry
rapidfuzz
pylint
boto3
python-dotenv
pandarallel